move createApiRoute to routes file
